requirements.txt
absl-py==0.9.0
alembic==1.4.2
astunparse==1.6.3
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
Flask==1.1.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gast==0.3.3
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
html5lib==1.0.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Preprocessing==1.1.0
lxml==4.5.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
MongoSanitizer==0.0.1
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
passlib==1.7.2
pdfminer==20191125
pdfminer.six==20200124
Pillow==7.0.0
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
pymongo==3.10.1
PyPDF2==1.26.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0
speaklater==1.3
SQLAlchemy==1.3.15
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post2
tensorflow==2.2.0rc2
tensorflow-estimator==2.2.0rc0
termcolor==1.1.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.0
wrapt==1.12.1
WTForms==2.2.1
